src/extract/requirements.txt
azure-storage-blob
ffmpeg-python